Split orders_spec in two: tests for orders list page and tests for orders edit page
